requirements.txt
pyspark==3.4.0
docker==7.1.0
click==8.1.7
PyYAML==6.0.1
